variotator-rest-service/requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.0
Jinja2==2.7.3
MarkupSafe==0.23
SPARQLWrapper==1.6.4
Werkzeug==0.9.6
aniso8601==0.90
argparse==1.2.1
gunicorn==19.1.1
html5lib==0.999
isodate==0.5.1
itsdangerous==0.24
pyenchant==1.6.6
pyparsing==2.0.3
pytz==2014.10
rdflib==4.1.2
six==1.8.0
wsgiref==0.1.2
